♻️ Split bootstrap and resources code in separate
