+++ test/vm/bootstrap/.env.sample
@@ -1,0 +1,7 @@
+IKNITE_BOOTSTRAP_REPO_URL=git@github.com:kaweezle/iknite.git
+IKNITE_BOOTSTRAP_REPO_REF=deploy/ama/local
+# The following format is used to construct the URL for fetching either the values or secrets of the project.
+# IKNITE_ENV_FORMAT="git::ssh://git@github.com/kaweezle/iknite.git@%s?ref=deploy/ama/local&cache=false"
+IKNITE_BOOTSTRAP_SCRIPT=iknite-bootstrap.sh
+GIT_SSH_COMMAND="ssh -i /workspace/.ssh/id_ed25519"
+SOPS_AGE_SSH_PRIVATE_KEY_FILE="/workspace/.ssh/id_ed25519"
